fix: cast X_L to float32 before cdist in create_attention_indices
bfloat16 distances have step size ~16 in the range 2048-4096 (angstroms
at high noise timesteps). This creates massive ties in D_LL, causing
topk's tie-breaking to behave non-deterministically and potentially
return non-unique indices, which then triggers the duplicate assertion
in get_sparse_attention_indices.

Casting X_L to float32 before cdist ensures stable, unique topk results
regardless of the AMP precision used for the model forward pass.

+++ models/rfd3/src/rfd3/model/layers/block_utils.py
@@ -196,7 +196,7 @@ def create_attention_indices(
         X_L = torch.randn(
             (1, L, 3), device=device, dtype=torch.float
         )  # [L, 3] - random
-    D_LL = torch.cdist(X_L, X_L, p=2)  # [B, L, L] - pairwise atom distances
+    D_LL = torch.cdist(X_L.float(), X_L.float(), p=2)  # [B, L, L] - always float32: bfloat16 quantises distances severely at large t, causing topk ties and duplicate attention indices
 
     # Create attention indices using neighbour distances
     base_mask = ~f["unindexing_pair_mask"][
